requirements.txt
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
POT==0.9.6.post1
Pillow==12.0.0
PyYAML==6.0.3
SQLAlchemy==2.0.44
Werkzeug==3.1.3
absl-py==2.3.1
alembic==1.17.1
ase==3.26.0
asttokens==3.0.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
cycler==0.12.1
dgl-cu111==0.6.1
dill==0.4.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
geometric==1.1
google-auth==2.43.0
google-auth-oauthlib==1.2.3
googledrivedownloader==1.1.0
greenlet==3.2.4
grpcio==1.76.0
h5py==3.15.1
huggingface-hub==1.1.2
icecream==2.1.8
idna==3.11
importlib-metadata==8.7.0
importlib-resources==6.5.2
isodate==0.7.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
littleutils==0.2.4
llvmlite==0.45.1
matplotlib==3.10.7
mkl-fft==2.1.1
mkl-random==1.3.0
mkl-service==2.6.0
multiprocess==0.70.18
networkx==3.5
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.10.2.21
oauthlib==3.3.1
ogb==1.3.6
optuna==4.6.0
outdated==0.2.2
packaging==25.0
pandas==2.3.3
pip==25.3
protobuf==6.33.0
psutil==7.1.3
pyaml==25.7.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2025.2
rdflib==7.4.0
rdkit==2025.9.1
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
ruamel.appconfig==0.5.5
ruamel.std.argparse==0.8.3
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torch-cluster==1.6.3
torch-geometric==2.7.0
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
urllib3==2.5.0
wheel==0.46.1
xxhash==3.6.0
yaml-1.3==0.1.0
zipp==3.23.0